+++ NEWS.md
@@ -1,7 +1,13 @@
 # pxmake (development version)
 
+## Breaking changes
+- Rename sheet 'Data' to 'Data table' and all internal functions from 
+'source_data' to 'data_table'.
+
 ## New features
 - Return rds object invisibly from `pxmake` and `metamake`.
+
+## Bug fixes and minor improvements
 
 # pxmake 0.7.0